Renames WPImmuTableCells > WPReusableTableViewCells
